// src/app/models/parametrosproductos.ts
export interface Parametrosproductos {
    PageNum?: number;
    PageSize?: number;
    tnCodigo?: String;       
    Detalle?: String;   
    IdArticulo?: number;
    IdMarca?: number;    
    IdCategoria?: number;
    IdSubCategoria?: number;
    IdFamlia?: number;
    tnTipoListado?: number;     
}